files: Move the files viewer and selection  bits to a separate widget
The motivation here is to break down the application into
smaller modules that are logically isolated and therefore
easier to maintain.

This change:

* Moves all the files viewing and selection UI components to a
  single PortfolioFiles widget.

* Keep the overall window nagivation and operations-activation
  logic to the existing PortfolioWindow widget.

Isolating files viewing and selection components to a separate
widget will also make it  easier to port this app to  GTK4 and
Libadwaita, later.

Issue #290
