+++ src/files.ui
@@ -1,0 +1,74 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated with glade 3.40.0 -->
+<interface>
+  <requires lib="gtk+" version="3.24"/>
+  <object class="GtkListStore" id="liststore">
+    <columns>
+      <!-- column-name icon -->
+      <column type="GIcon"/>
+      <!-- column-name name -->
+      <column type="gchararray"/>
+      <!-- column-name path -->
+      <column type="gchararray"/>
+    </columns>
+  </object>
+  <object class="GtkTreeModelFilter" id="filtered">
+    <property name="child-model">liststore</property>
+  </object>
+  <object class="GtkTreeModelSort" id="sorted">
+    <property name="model">filtered</property>
+  </object>
+  <template class="PortfolioFiles" parent="GtkScrolledWindow">
+    <property name="visible">True</property>
+    <property name="can-focus">True</property>
+    <property name="shadow-type">in</property>
+    <child>
+      <object class="GtkTreeView" id="treeview">
+        <property name="visible">True</property>
+        <property name="can-focus">True</property>
+        <property name="model">sorted</property>
+        <property name="headers-visible">False</property>
+        <property name="headers-clickable">False</property>
+        <property name="enable-search">False</property>
+        <property name="search-column">1</property>
+        <property name="fixed-height-mode">True</property>
+        <property name="show-expanders">False</property>
+        <property name="activate-on-single-click">True</property>
+        <child internal-child="selection">
+          <object class="GtkTreeSelection" id="selection"/>
+        </child>
+        <child>
+          <object class="GtkTreeViewColumn">
+            <property name="sizing">fixed</property>
+            <property name="title">icon</property>
+            <child>
+              <object class="GtkCellRendererPixbuf" id="icon">
+                <property name="width">50</property>
+                <property name="height">50</property>
+                <property name="stock_size">5</property>
+              </object>
+              <attributes>
+                <attribute name="gicon">0</attribute>
+              </attributes>
+            </child>
+          </object>
+        </child>
+        <child>
+          <object class="GtkTreeViewColumn" id="name_column">
+            <property name="sizing">fixed</property>
+            <property name="title">name</property>
+            <property name="expand">True</property>
+            <child>
+              <object class="GtkCellRendererText" id="name_cell">
+                <property name="ellipsize">end</property>
+              </object>
+              <attributes>
+                <attribute name="text">1</attribute>
+              </attributes>
+            </child>
+          </object>
+        </child>
+      </object>
+    </child>
+  </template>
+</interface>
