+++ src/window.ui
@@ -3,22 +3,6 @@
 <interface>
   <requires lib="gtk+" version="3.24"/>
   <requires lib="libhandy" version="1.0"/>
-  <object class="GtkListStore" id="liststore">
-    <columns>
-      <!-- column-name icon -->
-      <column type="GIcon"/>
-      <!-- column-name name -->
-      <column type="gchararray"/>
-      <!-- column-name path -->
-      <column type="gchararray"/>
-    </columns>
-  </object>
-  <object class="GtkTreeModelFilter" id="filtered">
-    <property name="child-model">liststore</property>
-  </object>
-  <object class="GtkTreeModelSort" id="sorted">
-    <property name="model">filtered</property>
-  </object>
   <object class="GtkPopover" id="menu_popover">
     <property name="can-focus">False</property>
     <child>
@@ -472,57 +456,12 @@
                                     <property name="visible">True</property>
                                     <property name="can-focus">False</property>
                                     <child>
-                                      <object class="GtkScrolledWindow" id="content_scroll">
+                                      <object class="GtkBox" id="content_inner_box">
                                         <property name="visible">True</property>
                                         <property name="can-focus">True</property>
-                                        <property name="shadow-type">in</property>
+                                        <property name="orientation">vertical</property>
                                         <child>
-                                          <object class="GtkTreeView" id="treeview">
-                                            <property name="visible">True</property>
-                                            <property name="can-focus">True</property>
-                                            <property name="model">sorted</property>
-                                            <property name="headers-visible">False</property>
-                                            <property name="headers-clickable">False</property>
-                                            <property name="enable-search">False</property>
-                                            <property name="search-column">1</property>
-                                            <property name="fixed-height-mode">True</property>
-                                            <property name="show-expanders">False</property>
-                                            <property name="activate-on-single-click">True</property>
-                                            <child internal-child="selection">
-                                              <object class="GtkTreeSelection" id="selection"/>
-                                            </child>
-                                            <child>
-                                              <object class="GtkTreeViewColumn">
-                                                <property name="sizing">fixed</property>
-                                                <property name="title">icon</property>
-                                                <child>
-                                                  <object class="GtkCellRendererPixbuf" id="icon">
-                                                    <property name="width">50</property>
-                                                    <property name="height">50</property>
-                                                    <property name="stock_size">5</property>
-                                                  </object>
-                                                  <attributes>
-                                                    <attribute name="gicon">0</attribute>
-                                                  </attributes>
-                                                </child>
-                                              </object>
-                                            </child>
-                                            <child>
-                                              <object class="GtkTreeViewColumn" id="name_column">
-                                                <property name="sizing">fixed</property>
-                                                <property name="title">name</property>
-                                                <property name="expand">True</property>
-                                                <child>
-                                                  <object class="GtkCellRendererText" id="name_cell">
-                                                    <property name="ellipsize">end</property>
-                                                  </object>
-                                                  <attributes>
-                                                    <attribute name="text">1</attribute>
-                                                  </attributes>
-                                                </child>
-                                              </object>
-                                            </child>
-                                          </object>
+                                          <placeholder/>
                                         </child>
                                       </object>
                                       <packing>
